# Optimize and rewrite sync::mpsc::unbounded