Unbounded queue is about 30% faster now in benchmarks.

Before the patch:

```
test bounded_100_tx   ... bench:     372,198 ns/iter (+/- 111,954)
test bounded_1_tx     ... bench:     387,090 ns/iter (+/- 108,744)
test unbounded_100_tx ... bench:     232,380 ns/iter (+/- 62,008)
test unbounded_1_tx   ... bench:     223,578 ns/iter (+/- 53,031)
```

With patch:

```
test bounded_100_tx   ... bench:     383,720 ns/iter (+/- 95,520)
test bounded_1_tx     ... bench:     372,090 ns/iter (+/- 105,511)
test unbounded_100_tx ... bench:     135,284 ns/iter (+/- 50,582)
test unbounded_1_tx   ... bench:     143,989 ns/iter (+/- 84,868)
```

The largest change is rewrite of `recv_task` with `AtomicPtr` instead
of `Mutex`.

Patch also separates unbounded and bounded queue implementations,
so unbounded queue no longer updates size of queue.

Patch adds `stress_drop_sender` test, because it helped to reveal
problems with stall at sender drop.

It is hard to prove that implementation does not contains concurrency bugs.
I've been running

```
while cargo test && cargo test --release; do; done
```

for the last 15 minutes, and so far so good.