@carllerche 2017-06-09
This is a pretty big PR and not really approachable w/o context. Just looking at it, it's not obvious to me why it is correct. Could you explain how the proposed changes are correct?