@stepancheg 2017-06-09
(Minor patch update, I accidentally changed semantics of `close` on `Sender` from no-op to real close, now reverted it back; writing now explanations, why patch is correct.)