@stepancheg 2017-06-09
While writing the explanation, I have found a bug. Sorry about that. Seems like it is not trivial to properly implement `receiver.close` without `num_message`. I'll reopen the PR later.